Move TemplateEntity to template (#98957)
* Move TemplateEntity to template

* Rename template_entity in helpers
